requirements.txt
scikit-learn==1.4.2
pandas
matplotlib
azureml-sdk
seaborn
fastapi 
uvicorn

seaborn
mlflow

python-dotenv
python-jose 
sqlalchemy
pytest
pytest-asyncio